User registration + spring version upgrade + (pom/maven) new dependency + change runtime to tomcat7 (compatibity : spring 4 & jsp api 2.2) + form bean validation

+++ SLM/src/main/java/jp/slm/business/dao/generic/GenericDao.java
@@ -18,9 +18,9 @@ public interface GenericDao<T, PK extends Serializable> {
 	/**
      * Count all.
      * 
-     * @return the integer
+     * @return the Long
      */
-	Integer countAll();
+	Long countAll();
 
 	/**
      * Count by criterion.
@@ -35,9 +35,9 @@ public interface GenericDao<T, PK extends Serializable> {
      * 
      * @param propertyName the property name
      * @param value the value
-     * @return the integer
+     * @return the Long
      */
-	Integer countByProperty(String propertyName, Object value);
+	Long countByProperty(String propertyName, Object value);
 
 	/**
      * Creates the.
